Jenkins job definition after several steps to evolve it
The job initially started by listing each program that was to be tested
as a separate program.  After several programs, that was boring and
repetitive.  The latest version uses a wildcard to select all programs
whose name started with the prefix "valgrind_".  The Makefile includes
a step to copy all programs to have a prefix valgrind_ in their name.

Still somewhat of a hassle not being able to do pre-execution and
post-execution steps, and not being able to pass arguments to the tested
program, but those should be possible to resolve with wrapper scripts.